+++ docs/architecture/sla_planner.md
@@ -29,7 +29,7 @@ Before using the SLA planner, you must profile the performance of the selected m
 
 ```bash
 cd $DYNAMO_HOME/benchmarks/profiler/
-python -m utils.profile_sla \
+python -m profile_sla \
   --config <path-to-dynamo-config-file> \
   --output-dir <path-to-profile-results-dir> \
   --isl <target-input-sequence-length> \
